distances to int

+++ src/main/java/org/newdawn/slick/util/pathfinding/AStarHeuristic.java
@@ -18,5 +18,5 @@ public interface AStarHeuristic {
 	 * be searched.
 
 	 */
-	public float getCost(TileBasedMap map, Mover mover, Point source, Point target);
+	public int getCost(TileBasedMap map, Mover mover, Point source, Point target);
 }
