+++ src/main/java/uk/co/gosseyn/xanax/domain/ForestZone.java
@@ -42,8 +42,8 @@ public class ForestZone extends Zone {
 
         return treeLocations.stream().sorted((o1, o2) -> {
             //TODO calc distance multiple times maybe?
-            float distance1 = o1.distanceTo(point);
-            float distance2 = o2.distanceTo(point);
+            int distance1 = o1.distanceTo(point);
+            int distance2 = o2.distanceTo(point);
             if(distance1 > distance2) {
                 return 1;
             } else if (distance1 < distance2) {
